+++ Source/RpmAvatarCreator/Private/RpmAvatarCreatorApi.cpp
@@ -34,14 +34,16 @@ URpmAvatarCreatorApi::URpmAvatarCreatorApi()
 	AvatarRequestHandler->ImageDownloader = ImageDownloader;
 }
 
-void URpmAvatarCreatorApi::SetPartnerDomain(const FString& PartnerDomain)
+void URpmAvatarCreatorApi::Initialize(const FString& Subdomain)
 {
-	AvatarProperties.Partner = PartnerDomain;
-	RequestFactory->SetPartnerDomain(PartnerDomain);
+	FString TrimmedSubdomain = Subdomain.TrimStartAndEnd();
+	checkf(!TrimmedSubdomain.IsEmpty(), TEXT("Application subdomain is required for the avatar creator. Find the subdomain of your application from the Ready Player Me studio website, and set it when constructing the avatar creator widget"));
+	AvatarProperties.Partner = TrimmedSubdomain;
+	RequestFactory->SetSubdomain(TrimmedSubdomain);
 	AuthManager->LoadUserData();
 	const UReadyPlayerMeSettings* Settings = GetDefault<UReadyPlayerMeSettings>();
 	const bool bIsAppIdSet = IsValid(Settings) && !Settings->AppId.IsEmpty();
-	checkf(bIsAppIdSet, TEXT("AppId is required for the avatar creator. Set your ReadyPlayerMe AppId in the project settings"));
+	checkf(bIsAppIdSet, TEXT("AppId is required for the avatar creator. Find the AppId of your application from the Ready Player Me studio website, and set it in your project settings under the ReadyPlayerMe > AppId"));
 }
 
 void URpmAvatarCreatorApi::SetPreviewDownloadedDelegate(const FPreviewDownloadCompleted& PreviewDownloaded)