+++ CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.0.0] 2023-09-01
+
+## Breaking
+- Partner Domain parameter is renamed to Subdomain in avatar creator widget.
+
+## Added
+- Support for modular outfits
+
+## Fixed
+- Assets are downloaded with parallel requests
+
 ## [0.5.0] 2023-08-24
 
 ## Breaking